fix(p24): psalm Batch 7 long tail, chunk 6 - final cluster, psalm clean
Real fixes: multi-line docblock union collapsed to one line
(DbConnection::params()), guard-then-bare-access extraction
(PwgCore date_min/date_max, PictureFormatsPageRenderer,
BatchManagerSubController), explicit string casts (QMultiToken
Stringable concat, CategoryRepository), in_array() replacing a
str_contains() heuristic false-positive, and a category-add repository
call site cleanup. Suppresses genuine PHPStan-vs-Psalm disagreements
(MessengerFactory's DoctrineTransportFactory generic resolution,
CategoryService's array_values() no-op per PHPStan's own list<>
tracking, ConfigurationSubController's already-proven-non-null truthy
checks, BrowserTestHelpers' ob_get_clean() stub gap) and pre-existing
deferred issues (PwgCategories return-shape narrowing, CommentService/
ConfigKeyExistsRule redundant guards) with documented rationale. Merges
a duplicate PossiblyNullPropertyAssignmentValue block.

composer analyse:psalm now exits 0.

+++ src/Piwigo/Controller/Admin/ConfigurationSubController.php
@@ -1049,7 +1049,7 @@ final class ConfigurationSubController implements AdminSubControllerInterface
                             [intval($pderivative['minw']), intval($pderivative['minh'])]
                         )
                     );
-                    $new_params->sharpen = intval($pderivative['sharpen']);
+                    $new_params->sharpen = (float) intval($pderivative['sharpen']);
 
                     ImageStdParams::apply_global($new_params);
